requirements.txt
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
docopt==0.4.0
Flask==2.3.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
mailchimp==2.0.10
MarkupSafe==3.0.2
pillow==11.3.0
python-dotenv==1.1.1
requests==2.32.4
urllib3==2.5.0
Werkzeug==2.3.8
mailchimp-marketing==3.0.0